+++ RptNetwork.cpp
@@ -28,7 +28,7 @@
 
 const unsigned int BUFFER_LENGTH = 200U;
 
-CRptNetwork::CRptNetwork(unsigned int localPort, const std::string& gwyAddress, unsigned int gwyPort, bool debug) :
+CRptNetwork::CRptNetwork(unsigned short localPort, const std::string& gwyAddress, unsigned short gwyPort, bool debug) :
 m_socket(localPort),
 m_addr(),
 m_addrLen(0U),
